Fix broken balance display in getinfo
- Move 'getinfo' method into wallet section of cli
